refactor: add emailSentEvent and emailService modify email verification time to last 3 minutes
