+++ petlog-api/src/main/java/com/ppp/api/email/handler/EmailSentEventHandler.java
@@ -1,0 +1,23 @@
+package com.ppp.api.email.handler;
+
+import com.ppp.api.email.dto.event.EmailSentEvent;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.event.TransactionPhase;
+import org.springframework.transaction.event.TransactionalEventListener;
+
+@Component
+@Slf4j
+@RequiredArgsConstructor
+public class EmailSentEventHandler {
+    private final JavaMailSender javaMailSender;
+
+    @Async
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void handleEmailSent(EmailSentEvent event) {
+        javaMailSender.send(event.getEmailCodeForm());
+    }
+}
